pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blame-tracker"
version = "0.1.0"
description = "Track code coverage gaps intersected with recent git changes"
requires-python = ">=3.8"
dependencies = [
    "pygments>=2.14.0",
    "jinja2>=3.1.0",
    "tqdm>=4.65.0",
    "GitPython>=3.1.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
blame-tracker = "blame_tracker.cli:main"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.ruff]
select = ["E", "F", "W", "I"]
line-length = 88
target-version = "py38"
